Update add, edit and delete functionality in general assessment

+++ src/pages/GeneralAssessment/index.js
@@ -91,20 +91,32 @@ export default () => {
         data.getGeneralData.edges.map(item => {
           if (item.node.modules && item.node.modules.edges.length > 0) {
             tempTable.push({
-              key: item.node.id,
+              id: item.node.id,
+              key: Math.random(),
               date: item.node.time,
               module: {
+                isParent: true,
                 pk: item.node.modules.edges[0].node.id,
                 id: item.node.modules.edges[0].node.module.id,
                 name: item.node.modules.edges[0].node.module.name,
                 score: item.node.modules.edges[0].node.score,
               },
-              note: item.node.note,
               submodules: item.node.submodules.edges.map(sbm => ({
                 pk: sbm.node.id,
                 id: sbm.node.submodule.id,
                 name: sbm.node.submodule.name,
                 score: sbm.node.score,
+              })),
+              note: item.node.note,
+              children: item.node.submodules.edges.map(sbm => ({
+                module: {
+                  pk: sbm.node.id,
+                  id: sbm.node.submodule.id,
+                  name: sbm.node.submodule.name,
+                  score: sbm.node.score,
+                },
+                id: sbm.node.submodule.id,
+                key: Math.random(),
               })),
             })
           }
@@ -128,7 +140,6 @@ export default () => {
     }
   }, [studentId])
 
-  console.log(data, loading, error, tableData, 'sd')
   console.log(currentRow)
   useEffect(() => {
     let tempList = originalData
@@ -199,7 +210,7 @@ export default () => {
       },
       sortOrder: sortedInfo.columnKey === 'date' && sortedInfo.order,
       sortDirections: ['ascend', 'descend'],
-      render: text => moment(text).format('YYYY-MM-DD'),
+      render: (text, row) => (row.module.isParent ? moment(text).format('YYYY-MM-DD') : null),
     },
     {
       title: 'Assessment',
@@ -221,32 +232,35 @@ export default () => {
       title: 'Actions',
       width: '250px',
       render: text => {
-        return (
-          <>
-            <Popconfirm
-              title="Are you sure you don't want this record?"
-              onConfirm={() => handleDelete(text)}
-              okText="Yes"
-              cancelText="No"
-            >
-              <Button type="link" style={{ color: 'red' }}>
-                <DeleteOutlined /> Delete
+        if (text.module.isParent) {
+          return (
+            <>
+              <Popconfirm
+                title="Are you sure you don't want this record?"
+                onConfirm={() => handleDelete(text)}
+                okText="Yes"
+                cancelText="No"
+              >
+                <Button type="link" style={{ color: 'red' }}>
+                  <DeleteOutlined /> Delete
+                </Button>
+              </Popconfirm>
+              <span style={{ borderRight: '1px solid #ccc', margin: '0 8px' }} />
+              <Button
+                type="link"
+                onClick={() => {
+                  setCurrentRow(text)
+                  setUpdate(true)
+                  setOpen(true)
+                }}
+              >
+                <Icon type="edit" />
+                Edit
               </Button>
-            </Popconfirm>
-            <span style={{ borderRight: '1px solid #ccc', margin: '0 8px' }} />
-            <Button
-              type="link"
-              onClick={() => {
-                setCurrentRow(text)
-                setUpdate(true)
-                setOpen(true)
-              }}
-            >
-              <Icon type="edit" />
-              Edit
-            </Button>
-          </>
-        )
+            </>
+          )
+        }
+        return null
       },
     },
   ]
